+++ visualization/app/codeCharta/ui/attributeSideBar/attributeSideBarPrimaryMetrics/attributeSideBarPrimaryMetrics.component.ts
@@ -4,7 +4,8 @@ import { Observable } from "rxjs"
 import { showAttributeTypeSelectorSelector } from "../util/showAttributeTypeSelector.selector"
 import { PrimaryMetrics, primaryMetricsSelector } from "../../../state/selectors/primaryMetrics/primaryMetrics.selector"
 import { Store } from "@ngrx/store"
-import { CcState } from "../../../codeCharta.model"
+import { AttributeDescriptors, CcState } from "../../../codeCharta.model"
+import { attributeDescriptorsSelector } from "../../../state/store/fileSettings/attributeDescriptors/attributeDescriptors.selector"
 
 @Component({
 	selector: "cc-attribute-side-bar-primary-metrics",
@@ -14,9 +15,11 @@ import { CcState } from "../../../codeCharta.model"
 export class AttributeSideBarPrimaryMetricsComponent {
 	primaryMetrics$: Observable<PrimaryMetrics>
 	showAttributeTypeSelector$: Observable<boolean>
+	attributeDescriptors$: Observable<AttributeDescriptors>
 
-	constructor(store: Store<CcState>) {
-		this.primaryMetrics$ = store.select(primaryMetricsSelector)
-		this.showAttributeTypeSelector$ = store.select(showAttributeTypeSelectorSelector)
+	constructor(private store: Store<CcState>) {
+		this.primaryMetrics$ = this.store.select(primaryMetricsSelector)
+		this.showAttributeTypeSelector$ = this.store.select(showAttributeTypeSelectorSelector)
+		this.attributeDescriptors$ = this.store.select(attributeDescriptorsSelector)
 	}
 }