+++ visualization/app/codeCharta/ui/metricChooser/metricChooser.component.ts
@@ -1,8 +1,8 @@
 import { Component, ElementRef, Input, OnInit, ViewChild, ViewEncapsulation } from "@angular/core"
 import { map, Observable } from "rxjs"
 import { EdgeMetricData, NodeMetricData, CcState } from "../../codeCharta.model"
-import { metricTitles } from "../../util/metric/metricTitles"
 import { metricDataSelector } from "../../state/selectors/accumulatedData/metricData/metricData.selector"
+import { attributeDescriptorsSelector } from "../../state/store/fileSettings/attributeDescriptors/attributeDescriptors.selector"
 import { Store } from "@ngrx/store"
 
 type MetricChooserType = "node" | "edge"
@@ -22,7 +22,7 @@ export class MetricChooserComponent implements OnInit {
 	@ViewChild("searchTermInput") searchTermInput: ElementRef<HTMLInputElement>
 	searchTerm = ""
 	metricData$: Observable<NodeMetricData[] | EdgeMetricData[]>
-	metricDataDescription: Map<string, string> = metricTitles
+	attributeDescriptors$ = this.store.select(attributeDescriptorsSelector)
 
 	constructor(private store: Store<CcState>) {}