+++ visualization/app/codeCharta/ui/ribbonBar/heightMetricChooser/heightMetricChooser.module.ts
@@ -2,7 +2,7 @@ import { CommonModule } from "@angular/common"
 import { NgModule } from "@angular/core"
 import { MetricChooserModule } from "../../metricChooser/metricChooser.module"
 import { MetricChooserValueHoveredModule } from "../../metricChooser/metricChooserValueHovered/metricChooserValueHovered.module"
-import { HeightMetricChooserComponent } from "./areaMetricChooser.component"
+import { HeightMetricChooserComponent } from "./heightMetricChooser.component"
 
 @NgModule({
 	imports: [CommonModule, MetricChooserModule, MetricChooserValueHoveredModule],